Move extension functions to namespace
