+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/StatusListTokenIntegrityValidator.kt
@@ -6,8 +6,8 @@ import at.asitplus.wallet.lib.cbor.VerifyCoseSignature
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureFun
 import at.asitplus.wallet.lib.data.MediaTypes
 import at.asitplus.wallet.lib.data.StatusListToken
+import at.asitplus.wallet.lib.extensions.ifFalse
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListTokenPayload
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.third_party.kotlin.ifFalse
 import at.asitplus.wallet.lib.jws.VerifyJwsObject
 import at.asitplus.wallet.lib.jws.VerifyJwsObjectFun
