+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/extensions/Boolean.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.data.rfc.tokenStatusList.third_party.kotlin
+package at.asitplus.wallet.lib.extensions
 
 fun Boolean.ifTrue(block: () -> Unit) {
     if (this) {
