+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/StatusListSerializationTest.kt
@@ -3,7 +3,7 @@ package at.asitplus.wallet.lib.data.rfc.tokenStatusList
 import at.asitplus.catchingUnwrapped
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusBitSize
-import at.asitplus.wallet.lib.toView
+import at.asitplus.wallet.lib.extensions.toView
 import io.kotest.assertions.throwables.shouldThrowAny
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.datatest.withData
