+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/StatusListTokenSerializationTest.kt
@@ -2,7 +2,7 @@ package at.asitplus.wallet.lib.data.rfc.tokenStatusList
 
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusBitSize
-import at.asitplus.wallet.lib.toView
+import at.asitplus.wallet.lib.extensions.toView
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
